move common methods to DeltaTransaction
